# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Swagger Doc For Blewglobal app",
        "version": "1.0.0",
        "title": "Swagger Blewglobal app",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/userlogin": {
            "post": {
                "tags": [
                    "User"
                ],
                "security": [{
                    "security_key": []
                }],
                "summary": "signUp",
                "description": "userlogin",
                "operationId": "userlogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userlogin"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
        "/editPost": {
            "post": {
                "tags": [
                    "User"
                ],
                "security": [{
                        "security_key": []
                    },
                    {
                        "auth_key": []
                    }
                ],
                "summary": "editPost",
                "description": "editPost",
                "operationId": "editPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/editPost"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/deletePost": {
            "post": {
                "tags": [
                    "User"
                ],
                "security": [{
                        "security_key": []
                    },
                    {
                        "auth_key": []
                    }
                ],
                "summary": "deletePost",
                "description": "deletePost",
                "operationId": "deletePost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/deletePost"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
        "/reportReasonsList": {
            "get": {
                "tags": [
                    "User"
                ],
                "security": [{
                    "security_key": []
                }, {
                    "auth_key": []
                }],
                "summary": "reportReasonsList",
                "description": "reportReasonsList",
                "operationId": "reportReasonsList",
                "responses": {}
            }
        },
...
    },
...
    "servers": [{
            "url": "http://localhost:3112/api"
        },
        {
            "url": "http://202.164.42.227:3112/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "auth_key": {
                "type": "apiKey",
                "name": "auth_key",
                "in": "header"
            },
            "security_key": {
                "type": "apiKey",
                "name": "security_key",
                "value": "blewglobal123",
                "in": "header"
            }
        },
        "schemas": {
...
            "deletePost": {
                "type": "object",
                "required": [],
                "properties": {
                    "postId": {
                        "type": "number"
                    }
                }
            },
            "editPost": {
                "type": "object",
                "required": [],
                "properties": {
                    "postId": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
...
            "userlogin": {
                "type": "object",
                "required": [],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}